Move classes in their own file
