+++ Sofa/framework/Simulation/Core/src/sofa/simulation/MappingGraph.h
@@ -32,14 +32,10 @@
  */
 
 #pragma once
-#include <sofa/core/BaseMapping.h>
-#include <sofa/core/behavior/BaseForceField.h>
-#include <sofa/core/behavior/BaseMass.h>
-#include <sofa/core/behavior/BaseMechanicalState.h>
-#include <sofa/simulation/Node.h>
 #include <sofa/simulation/config.h>
+#include <sofa/simulation/Node.h>
+#include <sofa/simulation/mappinggraph/ComponentGroupMappingGraphNode.h>
 
-#include <optional>
 #include <queue>
 
 namespace sofa::simulation
@@ -51,178 +47,6 @@ enum class VisitorApplication
     ALL_NODES,
     ONLY_MAPPED_NODES,
     ONLY_MAIN_NODES
-};
-
-/**
- * @brief Visitor interface for traversing the mapping graph.
- * 
- * This visitor pattern is used to allow algorithms to process nodes in a 
- * structured way without modifying their structure or coupling traversal logic 
- * with specific component types. Implementations must override visit methods 
- * corresponding to the behaviors they intend to handle.
- */
-class MappingGraphVisitor
-{
-public:
-    virtual ~MappingGraphVisitor() = default;
-
-    /**
-     * @brief Visits a mechanical state node.
-     * @param mstate The mechanical state component to visit.
-     */
-    virtual void visit(core::behavior::BaseMechanicalState&) {}
-    
-    /**
-     * @brief Visits a base mapping node.
-     * @param mapping The mapping component to visit.
-     */
-    virtual void visit(core::BaseMapping&) {}
-    
-    /**
-     * @brief Visits a force field behavior node.
-     * @param ff The force field component to visit.
-     */
-    virtual void visit(core::behavior::BaseForceField&) {}
-    
-    /**
-     * @brief Visits a mass behavior node.
-     * @param m The mass component to visit.
-     */
-    virtual void visit(core::behavior::BaseMass&) {}
-};
-
-/**
- * @brief Abstract base class for nodes in the mapping graph.
- * 
- * Provides common functionality for tracking prerequisites (parents) and dependencies 
- * (children), forming a Directed Acyclic Graph (DAG). It implements the Visitor pattern.
- */
-class BaseMappingGraphNode : public std::enable_shared_from_this<BaseMappingGraphNode>
-{
-public:
-    using SPtr = std::shared_ptr<BaseMappingGraphNode>;
-    friend class MappingGraph;
-
-    virtual ~BaseMappingGraphNode() = default;
-
-    /**
-     * @brief Accepts a visitor, allowing the graph to be processed by an external algorithm.
-     * @param visitor The concrete visitor implementation.
-     */
-    virtual void accept(MappingGraphVisitor& visitor) const = 0;
-
-    /**
-     * @brief Gets the name of the component represented by this node.
-     * @return The name string.
-     */
-    virtual std::string getName() const { return {}; }
-
-    enum class NodeType {
-        MechanicalState,
-        Mapping,
-        Component,
-        Group
-    };
-
-    virtual NodeType getType() const = 0;
-
-    /**
-     * @return True if the node has an ancestor which is a mapping node
-     */
-    bool isMapped() const;
-
-   private:
-    sofa::type::vector<SPtr> m_parents;   ///< prerequisite nodes (nodes pointing to this one)
-    sofa::type::vector<SPtr> m_children;  ///< dependent nodes (nodes pointed from this one)
-
-    // Mutable counter used during traversal (reset before each traversal).
-    mutable int m_pendingCount = 0;
-
-    mutable std::optional<bool> m_isMapped;
-};
-
-/**
- * @brief Template class representing a graph node associated with any SOFA component.
- * 
- * This wrapper allows the `MappingGraph` to manage nodes for various types of 
- * components (e.g., MechanicalState, ForceField) polymorphically while maintaining 
- * type safety and providing standard graph node interfaces.
- * 
- * @tparam TComponent The actual SOFA component class pointer type.
- */
-template<class TComponent>
-class MappingGraphNode : public BaseMappingGraphNode
-{
-public:
-    using SPtr = std::shared_ptr<MappingGraphNode>;
-    friend class MappingGraph;
-
-    /**
-     * @brief Constructs a node wrapper for the given component.
-     * @param s A shared pointer to the component instance.
-     */
-    explicit MappingGraphNode(typename TComponent::SPtr s)
-        : m_component(std::move(s))
-    {}
-
-    /**
-     * @brief Implements accept by calling visit on the wrapped component.
-     * @param visitor The concrete visitor implementation.
-     */
-    void accept(MappingGraphVisitor& visitor) const override
-    {
-        if (m_component)
-        {
-            visitor.visit(*m_component);
-        }
-    }
-
-    /**
-     * @brief Returns the name of the wrapped component.
-     * @return The name string from the component.
-     */
-    std::string getName() const override
-    {
-        return m_component->getName();
-    }
-
-    NodeType getType() const override
-    {
-        if constexpr (std::is_base_of_v<core::behavior::BaseMechanicalState, TComponent>)
-            return NodeType::MechanicalState;
-        else if constexpr (std::is_base_of_v<core::BaseMapping, TComponent>)
-            return NodeType::Mapping;
-        else
-            return NodeType::Component;
-    }
-
-   private:
-    typename TComponent::SPtr m_component; ///< The actual SOFA component instance pointer.
-};
-
-/**
- * @brief A node wrapper used for representing groups of components or abstract groupings.
- */
-class ComponentGroupMappingGraphNode : public BaseMappingGraphNode
-{
-public:
-    using SPtr = std::shared_ptr<ComponentGroupMappingGraphNode>;
-
-    void accept(MappingGraphVisitor& visitor) const override { SOFA_UNUSED(visitor); }
-
-    /**
-     * @brief Returns the fixed name "group" for this type of node.
-     * @return The string "group".
-     */
-    std::string getName() const override
-    {
-        return "group";
-    }
-
-    NodeType getType() const override
-    {
-        return NodeType::Group;
-    }
 };
 
 /**
