+++ Sofa/framework/Simulation/Core/src/sofa/simulation/mappinggraph/BaseMappingGraphNode.cpp
@@ -1,0 +1,43 @@
+/******************************************************************************
+*                 SOFA, Simulation Open-Framework Architecture                *
+*                    (c) 2006 INRIA, USTL, UJF, CNRS, MGH                     *
+*                                                                             *
+* This program is free software; you can redistribute it and/or modify it     *
+* under the terms of the GNU Lesser General Public License as published by    *
+* the Free Software Foundation; either version 2.1 of the License, or (at     *
+* your option) any later version.                                             *
+*                                                                             *
+* This program is distributed in the hope that it will be useful, but WITHOUT *
+* ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or       *
+* FITNESS FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License *
+* for more details.                                                           *
+*                                                                             *
+* You should have received a copy of the GNU Lesser General Public License    *
+* along with this program. If not, see <http://www.gnu.org/licenses/>.        *
+*******************************************************************************
+* Authors: The SOFA Team and external contributors (see Authors.txt)          *
+*                                                                             *
+* Contact information: contact@sofa-framework.org                             *
+******************************************************************************/
+#include <sofa/simulation/mappinggraph/BaseMappingGraphNode.h>
+
+namespace sofa::simulation
+{
+
+bool BaseMappingGraphNode::isMapped() const
+{
+    if (m_isMapped.has_value())
+    {
+        return m_isMapped.value();
+    }
+
+    const auto isMapped = std::any_of(m_parents.begin(), m_parents.end(), [](const SPtr& node)
+    {
+        return node->getType() == NodeType::Mapping || node->isMapped();
+    });
+
+    m_isMapped = isMapped;
+    return isMapped;
+}
+
+}
