+++ Sofa/framework/Simulation/Core/src/sofa/simulation/mappinggraph/MappingGraphVisitor.h
@@ -1,0 +1,71 @@
+/******************************************************************************
+*                 SOFA, Simulation Open-Framework Architecture                *
+*                    (c) 2006 INRIA, USTL, UJF, CNRS, MGH                     *
+*                                                                             *
+* This program is free software; you can redistribute it and/or modify it     *
+* under the terms of the GNU Lesser General Public License as published by    *
+* the Free Software Foundation; either version 2.1 of the License, or (at     *
+* your option) any later version.                                             *
+*                                                                             *
+* This program is distributed in the hope that it will be useful, but WITHOUT *
+* ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or       *
+* FITNESS FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License *
+* for more details.                                                           *
+*                                                                             *
+* You should have received a copy of the GNU Lesser General Public License    *
+* along with this program. If not, see <http://www.gnu.org/licenses/>.        *
+*******************************************************************************
+* Authors: The SOFA Team and external contributors (see Authors.txt)          *
+*                                                                             *
+* Contact information: contact@sofa-framework.org                             *
+******************************************************************************/
+#pragma once
+
+#include <sofa/simulation/config.h>
+#include <sofa/core/BaseMapping.h>
+#include <sofa/core/behavior/BaseForceField.h>
+#include <sofa/core/behavior/BaseMass.h>
+#include <sofa/core/behavior/BaseMechanicalState.h>
+
+namespace sofa::simulation
+{
+
+/**
+ * @brief Visitor interface for traversing the mapping graph.
+ *
+ * This visitor pattern is used to allow algorithms to process nodes in a
+ * structured way without modifying their structure or coupling traversal logic
+ * with specific component types. Implementations must override visit methods
+ * corresponding to the behaviors they intend to handle.
+ */
+class MappingGraphVisitor
+{
+public:
+    virtual ~MappingGraphVisitor() = default;
+
+    /**
+     * @brief Visits a mechanical state node.
+     * @param mstate The mechanical state component to visit.
+     */
+    virtual void visit(core::behavior::BaseMechanicalState&) {}
+
+    /**
+     * @brief Visits a base mapping node.
+     * @param mapping The mapping component to visit.
+     */
+    virtual void visit(core::BaseMapping&) {}
+
+    /**
+     * @brief Visits a force field behavior node.
+     * @param ff The force field component to visit.
+     */
+    virtual void visit(core::behavior::BaseForceField&) {}
+
+    /**
+     * @brief Visits a mass behavior node.
+     * @param m The mass component to visit.
+     */
+    virtual void visit(core::behavior::BaseMass&) {}
+};
+
+}
